+++ src/DataValidator/Filters/ExistsInDB.php
@@ -36,13 +36,13 @@ class ExistsInDB extends AbstractFilter
 
     /**
      * Constructor
-     * @since 1.0
+     * @since 2.1
      * @param string $table Database table name
      * @param string $field Database field name
-     * @param string $message Error message (Defaults to NULL)
      * @param DatabaseConnectionInstance $connection Database connection instance
+     * @param string $message Error message (Defaults to NULL)
      */
-    public function __construct(string $table, string $field, ?string $message = null, DatabaseConnectionInstance $connection)
+    public function __construct(string $table, string $field, DatabaseConnectionInstance $connection, ?string $message = null)
     {
         if (empty($message)) {
             $message = sprintf(dgettext('mfx', "The '%%s' field must reprensent an existing entry in the '%s' table (matched on the '%s' field)."), $table, $field);